pyproject.toml
[tool.poetry]
name = "os_downloader_flask"
version = "0.1.0"
description = ""
authors = ["Franco N. Merenda <franco.merenda@gordiansoftware.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
bidict = "^0.22.1"
cachelib = "^0.10.0"
click = "^8.1.3"
Flask-Login = "^0.6.2"
Flask-Session = "^0.4.0"
Flask-SocketIO = "^5.3.2"
importlib-metadata = "^6.0.0"
itsdangerous = "^2.1.2"
python-socketio = "^5.7.2"
python-engineio = "^4.3.4"
six = "^1.16.0"
Werkzeug = "^2.2.2"
zipp = "^3.11.0"
requests = "^2.28.2"
celery = "^5.2.7"
redis = "^4.4.2"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
